+++ vite-project/package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0",
     "styled-components": "^6.1.0"
   },
   "devDependencies": {
