+++ vite-project/src/main.tsx
@@ -5,5 +5,5 @@ import { App } from "./App.tsx";
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
+  </React.StrictMode>
 );
